refactor: split useTimelinePlayer.ts and hyperframes-player.ts into focused modules (<500 LOC each)
